search engine implementation

+++ searchengine-master/src/main/java/searchengine/repository/IndexRepository.java
@@ -3,7 +3,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import searchengine.entity.Index;
 
-public interface IndexRepository extends JpaRepository<Index, Long> {
+public interface IndexRepository extends JpaRepository<Index, Integer> {
     Index findByPageIdAndLemma(Integer pageId, String lemma);
 }
